refactor: mod {engine -> enclave_use_case}
