+++ frame/enclave/src/lib.rs
@@ -1,7 +1,7 @@
 #![no_std]
 extern crate sgx_tstd as std;
 
-pub mod engine;
+pub mod enclave_use_case;
 mod register_enclave_use_case;
 
-pub use crate::engine::*;
+pub use crate::enclave_use_case::*;
